Refactor to publish common middlewares
