+++ lib/fiware-iotagent-lib.js
@@ -27,6 +27,7 @@ var async = require('async'),
     apply = async.apply,
     ngsi = require('./services/ngsi/ngsiService'),
     intoTrans = require('./services/common/domain').intoTrans,
+    middlewares = require('./services/common/genericMiddleware'),
     db = require('./model/dbConn'),
     subscriptions = require('./services/ngsi/subscriptionService'),
     statsRegistry = require('./services/stats/statsRegistry'),
@@ -198,6 +199,9 @@ exports.intoTrans = intoTrans;
 
 exports.ensureSouthboundDomain = domainUtils.ensureSouthboundDomain;
 exports.finishSouthBoundTransaction = domainUtils.finishSouthBoundTransaction;
+exports.requestDomain = domainUtils.requestDomain;
+
+exports.middlewares = middlewares;
 
 exports.commandLine = require('./command/commandLine');
