+++ lib/services/northBound/contextServer.js
@@ -30,7 +30,7 @@ var async = require('async'),
     ngsi = require('../ngsi/ngsiService'),
     intoTrans = require('../common/domain').intoTrans,
     deviceService = require('../devices/deviceService'),
-    revalidator = require('revalidator'),
+    middlewares = require('../common/genericMiddleware'),
     _ = require('underscore'),
     context = {
         op: 'IoTAgentNGSI.ContextServer'
@@ -622,32 +622,6 @@ function setNotificationHandler(newHandler) {
     notificationHandler = newHandler;
 }
 
-
-function ensureType(req, res, next) {
-    if (req.is('json') || req.is('xml')) {
-        next();
-    } else {
-        next(new errors.UnsupportedContentType(req.headers['content-type']));
-    }
-}
-
-function validateJson(template) {
-    return function validate(req, res, next) {
-        if (req.is('json')) {
-            var errorList = revalidator.validate(req.body, template);
-
-            if (errorList.valid) {
-                next();
-            } else {
-                logger.debug('Errors found validating request: %j', errorList);
-                next(new errors.BadRequest('Errors found validating request.'));
-            }
-        } else {
-            next();
-        }
-    };
-}
-
 /**
  * Load the templates that will be used to generate XML Payloads for Orion.
  */
@@ -683,15 +657,15 @@ function loadTemplates() {
 function loadContextRoutes(router) {
     //TODO: remove '//' paths when the appropriate patch comes to Orion
     var updateMiddlewares = [
-            ensureType,
+            middlewares.ensureType,
             ngsiParser.readUpdateBody,
-            validateJson(updateContextTemplate),
+            middlewares.validateJson(updateContextTemplate),
             handleUpdate
         ],
         queryMiddlewares = [
-            ensureType,
+            middlewares.ensureType,
             ngsiParser.readQueryBody,
-            validateJson(queryContextTemplate),
+            middlewares.validateJson(queryContextTemplate),
             handleQuery
         ],
         updatePaths = [
@@ -713,8 +687,8 @@ function loadContextRoutes(router) {
     }
 
     router.post('/notify', [
-        ensureType,
-        validateJson(notificationTemplate),
+        middlewares.ensureType,
+        middlewares.validateJson(notificationTemplate),
         handleNotification
     ]);
