+++ lib/services/northBound/northboundServer.js
@@ -28,69 +28,20 @@ var http = require('http'),
     northboundServer,
     contextServer = require('./contextServer'),
     domainUtils = require('../common/domain'),
+    middlewares = require('../common/genericMiddleware'),
     intoTrans = domainUtils.intoTrans,
     deviceProvisioning = require('./deviceProvisioningServer'),
     groupProvisioning = require('./deviceGroupAdministrationServer'),
     logger = require('logops'),
     utils = require('./restUtils'),
-    iotaInformation,
     context = {
         op: 'IoTAgentNGSI.NorthboundServer'
     },
     bodyParser = require('body-parser');
 
-function handleError(error, req, res, next) {
-    var code = 500;
-
-    logger.debug('Error [%s] handing request: %s', error.name, error.message);
-
-    if (error.code && String(error.code).match(/^[2345]\d\d$/)) {
-        code = error.code;
-    }
-
-    res.status(code).json({
-        name: error.name,
-        message: error.message
-    });
-}
-
-function traceRequest(req, res, next) {
-    logger.debug('Request for path [%s] from [%s]', req.path, req.get('host'));
-
-    if (req.is('json')) {
-        logger.debug('Body:\n\n%s\n\n', JSON.stringify(req.body, null, 4));
-    } else if (req.is('xml')) {
-        logger.debug('Body:\n\n%s\n\n', req.rawBody);
-    } else {
-        logger.debug('Unrecognized body type', req.headers['content-type']);
-    }
-
-    next();
-}
-
-function retrieveVersion(req, res, next) {
-    res.status(200).json(iotaInformation);
-}
-
-function changeLogLevel(req, res, next) {
-    var levels = ['INFO', 'ERROR', 'FATAL', 'DEBUG', 'WARNING'];
-
-    if (!req.query.level) {
-        res.status(400).json({
-            error: 'log level missing'
-        });
-    } else if (levels.indexOf(req.query.level.toUpperCase()) < 0) {
-        res.status(400).json({
-            error: 'invalid log level'
-        });
-    } else {
-        logger.setLevel(req.query.level.toUpperCase());
-        res.status(200).send('');
-    }
-}
-
 function start(config, callback) {
-    var baseRoot = '/';
+    var baseRoot = '/',
+        iotaInformation;
 
     northboundServer = {
         server: null,
@@ -108,7 +59,7 @@ function start(config, callback) {
     northboundServer.app.use(bodyParser.json());
 
     if (config.logLevel && config.logLevel === 'DEBUG') {
-        northboundServer.app.use(traceRequest);
+        northboundServer.app.use(middlewares.traceRequest);
     }
 
     if (config.server.baseRoot) {
@@ -125,16 +76,18 @@ function start(config, callback) {
         iotaInformation.version = config.iotaVersion;
     }
 
-    northboundServer.router.get('/iot/about', retrieveVersion);
-    northboundServer.router.get('/version', retrieveVersion);
-    northboundServer.router.put('/admin/log', changeLogLevel);
+    middlewares.setIotaInformation(iotaInformation);
+
+    northboundServer.router.get('/iot/about', middlewares.retrieveVersion);
+    northboundServer.router.get('/version', middlewares.retrieveVersion);
+    northboundServer.router.put('/admin/log', middlewares.changeLogLevel);
 
     northboundServer.app.use(baseRoot, northboundServer.router);
     contextServer.loadContextRoutes(northboundServer.router);
     deviceProvisioning.loadContextRoutes(northboundServer.router);
     groupProvisioning.loadContextRoutes(northboundServer.router);
 
-    northboundServer.app.use(handleError);
+    northboundServer.app.use(middlewares.handleError);
 
     northboundServer.server = http.createServer(northboundServer.app);
